Added overflow check in normalsDirectionCount
Former-commit-id: e1a3fedd4e1e174d4df6e50b47ae624b1a7a87f0

+++ modules/odm_25dmeshing/src/PlyInterpreter.hpp
@@ -3,6 +3,7 @@
 #include <utility>
 #include <vector>
 #include <fstream>
+#include <limits>
 
 #include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
 #include <CGAL/property_map.h>
@@ -19,7 +20,7 @@ typedef std::pair<Point3, Vector3> Pwn;
 
 class PlyInterpreter {
 	std::vector<Pwn>& points;
-	int zNormalsDirectionCount;
+	long zNormalsDirectionCount;
 
 	public:
 	 PlyInterpreter (std::vector<Pwn>& points)
